Okay.
I took a look at how I wanted to store miner stats, and I think I'm going to move forward with a separate class that stores all the miner stats for a particular pool instance.

It will be handled by the MinerManager for that pool. This will allow miners to come and go, but store their stats based on their RigName, possibly for use in a Web UI or other app in the future.